extract Env into a separate module and draft the implementation;
