+++ ergo-lib/src/ast/bin_op.rs
@@ -3,7 +3,7 @@
 use eval::costs::Costs;
 
 use crate::eval;
-use crate::eval::Env;
+use crate::eval::env::Env;
 use crate::eval::EvalContext;
 use crate::eval::EvalError;
 use crate::eval::Evaluable;
