+++ ergo-lib/src/sigma_protocol/prover.rs
@@ -19,7 +19,8 @@ use super::{
 };
 use crate::ergo_tree::{ErgoTree, ErgoTreeParsingError};
 use crate::eval::context::Context;
-use crate::eval::{Env, EvalError, Evaluator};
+use crate::eval::env::Env;
+use crate::eval::{EvalError, Evaluator};
 use thiserror::Error;
 
 /// Prover errors
